+++ AlbaGo/src/main/java/com/albago/domain/PayVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class PayVO {
 	private int p_code;
-	private int c_code;
+	private long c_code;
 	private String u_id;
 	private int p_perhour;
 }
